requirements.txt
discord.py
discord-flags
python-dotenv